refactor(accounts): split api.py into api/ package
Move auth, signup, and profile endpoints into apps/accounts/api/*,
with WorkOS lookup/refresh helpers in api/workos_match.py. The former
api.py becomes api/__init__.py, a thin facade exporting the routers
list. Matches the existing apps/catalog/api/ convention.

Drain the inline-schema baseline for apps.accounts.api from 4 to 0:
- AuthStatusSchema moves into the shared apps/accounts/schemas.py.
- EntityContributionSchema, UserChangeSetSchema, UserProfileSchema
  move into api/profile_schemas.py, scoped to the user-profile page
  endpoint. Keeping them out of the shared accounts schema module
  prevents the accounts->provenance dependency (via ChangeSet) from
  leaking into auth/signup, which only the page-endpoint carve-out
  in AppBoundaries.md permits.

Add 0-baselines for api.{auth,profile,signup} so the inline-schema
tripwire keeps holding against the new modules.
